Introduce cliff for cli framework
This patch introduces the cliff framework from oslo.
The same functionality from the old command line interface was
replicated, and a few orthographic errors where corrected from the
command help.

The application structure was made so that it resembles the way
python-openstackclient was done, this will enable the code to be
integrated to that client in a more straight forward way. Same with
the naming of classes within the sub-commands. For example, the list
commands are List<Entity> and not List<Entities>; this was done in
such way because it is done that way in python-openstack client;
and the aim of this commit is to produce functional code, and at the
same time, follow their standards to be able to integrate this
there without much hassle.

NOTE: Only secrets and orders were added, since verifications are no
      longer in use

Implements: blueprint cliff-for-python-barbicanclient
Change-Id: Ice2dddf418dfb76a616b65f22dc8dfd7ef4df36f