Enable passing test regex to testr from tox
This will enable us to pass a regex to tox (such as we already can for
testr) of what tests we want to execute. So we can do stuff like this:

$ tox -e py27,py33 -- test_barbican

Which will execute the tests that match that "test_barbican" argument
(which will actually be the tests in the test_barbican module) on the
given environments.

Change-Id: Id81b8ab4829c25454ee049b22e3e13a551055946