Enable usage of oslo_debug_helper from tox
The script oslo_debug_helper enables the usage of Python's pdb without
having to do the tedious listing of test cases from testr and then
running it. Only thing that needs to be done now is the adding of
"import pdb; pdb.set_trace()" as usual and then one can just call:

    $ tox -e debug

And we will end up in the debugger when that test is hit.

Note that you can also give a test_regex such as:

    $ tox -e debug -- some_test_regex

To run only tests that match that regex.

Change-Id: I63386e68ac3ef5c05bc8290d889efa7f71b33cf1