Fixing the broken functional tests
Adding creator_id for Secret and Order types
Adding sub_status and sub_status_message for Order types
This is a result of these attributes being added to the API.
This change isn't aimed to support these fields, but rather just
accepts the data so that it won't bork the tests. Actual support
for using these fields should be addressed in a follow-up patch.

Fixes bug 1441848
Change-Id: I2340b7dd1dae3806418fa99bdaad73860dda5fbc