Use keystoneclient to get endpoint if it's empty
We used to rely on keystoneclient code to get the proper endpoint from
keystone if it was not provided in the command line or as an
environment variable. This changed in a recent commit. So now that we
are using an adapter for the keystone sessions, we use the function
provided by that adapter.

Change-Id: I927a105831642d4fb00691ea2a97e27e79ebc7fd
Closes-Bug: #1443436