Remove py26 support
As of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.
Closes-Bug: 1519510
Depends-On: Ic8241e55b5f4e3d567156d4c1e8bf8a0303f6d29

Change-Id: I4386dc70e3d467e31963d852726af78fbdbbba38