Remove invalid smoke test that is failing in the gate
The smoke test that ran the CLI command "acl delete"
with a null entity_ref was failing with an exception
in the cliff library.  This test is invalid, since
the CLI parser would not accept this command without
an entity_ref.

This test failed more gracefully with cliff 2.3.0, but
in 2.4.0 cliff throws an exception that causes this
test to start failing.

Change-Id: If91d4966eff77dad7c8569e309134dc540edc258