Add --file flag for secrets
This patch adds a --file (-F) flag to both 'secret store' and 'secret
get' commands to work around a cliff formatter limitation when dealing
with binary data. This flag can be used to store a payload directly
from a file so that a user doesn't have to do any base64 juggling of
their data.

The flag can also be used to save a payload directly to a file so we can
avoid UTF-8 and ASCII encoding errors thrown by the cliff formatters
that expect text data.

Original Review: https://review.openstack.org/#/c/388981
Co-Authored-By: douglas.mendizabal@rackspace.com

Change-Id: Iee7e2bbb95238a9bbb9e8e6124fe1663da377939