Pass OSC interface through to barbican plugin
python-openstackclient uses OS_INTERFACE or --os-interface to select a
catalog endpoint to work with. The barbican CLI uses BARBICAN_INTERFACE
or --interface, which OSC won't pass through. This is a problem because
the it prevents the user from selecting a non-public endpoint when using
barbican via openstackclient rather than as a standalone CLI. This patch
updates the osc plugin to pass through the interface set by OSC to the
barbican client class.

Change-Id: I1e9890c9e87fb77b272f939bf33914df739aa5be