Secret payload should also be fetched by UUID
We changed the client to fetch containers and secrets via their UUID
from the API, rather than by HREF, so that the endpoint URLs set in the
keystone client would be respected. Unfortunately, we (I) missed
updating the payload fetch function to do the same. This brings it into
line with the other fetches.

Change-Id: Ic71cf6771563d669a2fa37a56d4b40c637db1511
Story: 2004653
Task: 28608