Drop lower-constraints.txt and its testing
As discussed in TC PTG[1] and TC resolution[2], we are
dropping the lower-constraints.txt file and its testing.
We will keep lower bounds in the requirements.txt file but
with a note that these are not tested lower bounds and we
try our best to keep them updated.

This patch also fixes a broken test by updating the maximum
payload size to match the current maximum in Barbican. [3]

This patch also skips an additional broken test.  It will be
fixed in a separate patch.

[1] https://etherpad.opendev.org/p/tc-zed-ptg#L326
[2] https://governance.openstack.org/tc/resolutions/20220414-drop-lower-constraints.html#proposal
[3] https://opendev.org/openstack/barbican/src/branch/stable/yoga/barbican/common/config.py#L33

Change-Id: Id192b8eef2c2518b8a89f200727bd03c47dbd07e
(cherry picked from commit 8e3a391fd63c0a319071fc55996faab3e6699a0f)