Fix functional tests
The 'skip_sdist' option in 'tox.ini' changes behavior in tox 4. In tox
3, it was ignored if 'usedevelop = true' was specified. In tox 4, the
opposite is true. This means our plugin was not installed in the
environment. Fix this.

While we're here, cleanup the whole 'tox.ini' file and remove a lot of
stuff that is no longer necessary in a world where everything is Python
3.

Change-Id: Iee4ef8e1c15e67c4ab9a77676214f1062bd0c96a
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>